grouping to error.ts
