--- src/error/not-expect.ts
@@ -1,6 +1,0 @@
-export class NotExpectError extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = 'NotExpectError';
-    }
-}
