--- src/error/not-found.ts
@@ -1,6 +1,0 @@
-export class NotFoundError extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = 'NotFoundError';
-    }
-}
