--- src/error/not-ready.ts
@@ -1,6 +1,0 @@
-export class NotReadyError extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = 'NotReadyError';
-    }
-}
