--- src/error/not-setted.ts
@@ -1,6 +1,0 @@
-export class NotSettedError extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = 'NotSettedError';
-    }
-}
